Telbot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
frozenlist==1.3.3
magic-filter==1.0.9
multidict==6.0.4
pip==21.1.2
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
setuptools==57.0.0
six==1.16.0
telebot==0.0.5
urllib3==1.26.15
wheel==0.36.2
yarl==1.8.2
python-dotenv==1.0.0
